+++ SR2MP/Server/Handlers/PlayerJoinHandler.cs
@@ -34,7 +34,7 @@ public sealed class PlayerJoinHandler : BasePacketHandler
 
         var joinPacket = new PlayerJoinPacket
         {
-            Type = (byte)PacketType.BroadcastPlayerJoin,
+            Type = PacketType.BroadcastPlayerJoin,
             PlayerId = playerId,
             PlayerName = packet.PlayerName
         };
